Moved the soundNotification to the right place
